complete task

+++ 0x06-pointers_arrays_strings/7-leet.c
@@ -14,7 +14,7 @@ char *leet(char *str)
 	char *s1 = "aAeEoOtTlL";
 	char *s2 = "4433007711";
 
-	for (i = 0; i < strlen(str); i++)
+	for (i = 0; i < (int) strlen(str); i++)
 	{
 		for (j = 0; j < 10; j++)
 		{
